feat: Migrate profile store to TypeScript
- Replaced the existing JavaScript implementation of the profile store with a TypeScript version in `profile.ts`.
- Added type definitions for user and queue permissions to enhance type safety and improve code maintainability.
- Retained existing functionality for managing user profile data and queues.
